refactor embedded -> local
