+++ cmd/bolted/ls/command.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/draganm/bolted"
 	"github.com/draganm/bolted/dbpath"
-	"github.com/draganm/bolted/embedded"
+	"github.com/draganm/bolted/local"
 	"github.com/urfave/cli/v2"
 	"go.etcd.io/bbolt"
 )
@@ -40,7 +40,7 @@ var Command = &cli.Command{
 			return fmt.Errorf("while parsing path %s: %w", p, err)
 		}
 
-		db, err := embedded.Open(sourceFile, 0700, embedded.Options{
+		db, err := local.Open(sourceFile, 0700, local.Options{
 			Options: bbolt.Options{
 				Timeout:  c.Duration("open-timeout"),
 				ReadOnly: true,
