+++ local/observer_test.go
@@ -1,4 +1,4 @@
-package embedded_test
+package local_test
 
 import (
 	"testing"
@@ -6,13 +6,13 @@ import (
 
 	"github.com/draganm/bolted"
 	"github.com/draganm/bolted/dbpath"
-	"github.com/draganm/bolted/embedded"
+	"github.com/draganm/bolted/local"
 	"github.com/stretchr/testify/require"
 )
 
 func TestObservePath(t *testing.T) {
 
-	db, cleanupDatabase := openEmptyDatabase(t, embedded.Options{})
+	db, cleanupDatabase := openEmptyDatabase(t, local.Options{})
 	defer cleanupDatabase()
 
 	updates, close := db.Observe(dbpath.ToPath("foo").ToMatcher().AppendAnySubpathMatcher())
